+++ level/unit/src/main/java/org/testifyproject/level/unit/UnitTestRunner.java
@@ -21,6 +21,7 @@ import org.testifyproject.TestContext;
 import org.testifyproject.TestDescriptor;
 import org.testifyproject.TestRunner;
 import org.testifyproject.core.util.ServiceLocatorUtil;
+import org.testifyproject.extension.CollaboratorReifier;
 import org.testifyproject.extension.FinalReifier;
 import org.testifyproject.extension.InitialReifier;
 import org.testifyproject.extension.PostVerifier;
@@ -29,7 +30,6 @@ import org.testifyproject.extension.PreiVerifier;
 import org.testifyproject.extension.SutReifier;
 import org.testifyproject.extension.annotation.UnitCategory;
 import org.testifyproject.tools.Discoverable;
-import org.testifyproject.extension.CollaboratorReifier;
 
 /**
  * A class used to run a integration test.